+++ axum-macros/tests/typed_path/fail/missing_capture.rs
@@ -1,0 +1,10 @@
+use axum_macros::TypedPath;
+use serde::Deserialize;
+
+#[derive(TypedPath, Deserialize)]
+#[typed_path("/users")]
+struct MyPath {
+    id: u32,
+}
+
+fn main() {}