+++ axum-macros/tests/typed_path/fail/not_deserialize.rs
@@ -1,0 +1,9 @@
+use axum_macros::TypedPath;
+
+#[derive(TypedPath)]
+#[typed_path("/users/:id")]
+struct MyPath {
+    id: u32,
+}
+
+fn main() {}